Pronic Number without Loop
Got to pronic number without loop by first creating a check to see if the last digit was 0, 2, or 6. If it was not one of those then it was not a pronic number, then I took the square root of the number and took the floor of it which gave the x for the formula. If the result was not equal to the number entered then the result could not be a pronic number.